requirements.txt
Flask==2.2.2
Flask_WTF==1.0.1
google_api_python_client==2.76.0
keras==2.11.0
matplotlib==3.5.1
nltk==3.5
pandas==1.4.0
python-dotenv==0.21.1
regex==2022.6.2
scikit_learn==1.2.1
seaborn==0.11.2
tensorflow==2.11.0
tensorflow_hub==0.12.0
tensorflow_intel==2.11.0
gunicorn==20.1.0
psycopg2-binary==2.9.4
regex==2022.6.2
WTForms==3.0.1
